.js to .ts(tsx) for notifications / Switch
