+++ src/qryn-ui/Switch/Switch.styles.tsx
@@ -1,8 +1,8 @@
 import styled from "@emotion/styled";
 
 const Label = styled.div`
-    color: ${(props) => props.theme.textColor};
-    background: ${(props) => props.theme.buttonInactive};
+    color: ${(props: any) => props.theme.textColor};
+    background: ${(props: any) => props.theme.buttonInactive};
     display: flex;
     align-items: center;
     justify-content: center;
@@ -14,31 +14,31 @@ const SwitchCont = styled.div`
     display: flex;
     align-items: center;
     font-size: 12px;
-    background-color: ${(props) => props.theme.buttonInactive};
+    background-color: ${(props: any) => props.theme.buttonInactive};
 
-    border: 1px solid ${(props) => props.theme.buttonBorder};
-    color: ${(props) => props.theme.textColor};
+    border: 1px solid ${(props: any) => props.theme.buttonBorder};
+    color: ${(props: any) => props.theme.textColor};
     border-radius: 3px;
     margin-right: 10px;
 `;
 
-const SwitchBtn = styled.div`
+const SwitchBtn: any = styled.div`
     cursor: pointer;
     display: flex;
     align-items: center;
-    background: ${(props) =>
+    background: ${(props: any) =>
         props.selected
             ? props.theme.buttonDefault
             : props.theme.buttonInactive};
-    border-left: ${(props) =>
+    border-left: ${(props: any) =>
         props.position === "last"
             ? `1px solid ${props.theme.buttonBorder}`
             : "none"};
-    border-right: ${(props) =>
+    border-right: ${(props: any) =>
         props.position === "first"
             ? `1px solid ${props.theme.buttonBorder}`
             : "none"};
-    border-radius: ${({ position }) =>
+    border-radius: ${({ position }: any) =>
         position === "first"
             ? "3px 0px 0px 3px"
             : position === "last"
@@ -56,5 +56,4 @@ export const styles = {
     Label,
     SwitchCont,
     SwitchBtn,
-    
 }
